+++ template/frontend/app/app.config.ts
