+++ template/frontend/app/pages/index.vue
