+++ template/frontend/{% if frontend_uses_graphql %}codegen.ts{% endif %}
@@ -2,9 +2,9 @@ import type { CodegenConfig } from "@graphql-codegen/cli";
 
 const config: CodegenConfig = {
   schema: process.env.GRAPHQL_ENDPOINT || "http://127.0.0.1:4000/api/graphql",
-  documents: "./graphql/queries/**/*.gql",
+  documents: "./app/graphql/queries/**/*.gql",
   generates: {
-    "./graphql/generated/graphql.ts": {
+    "./app/graphql/generated/graphql.ts": {
       plugins: ["typescript", "typescript-operations", "typescript-vue-apollo"], // cannot use `preset: "client"` because it causes duplications with typescript-vue-apollo
       config: {
         documentMode: "graphQLTag",
